- cleanup code - fix signed/unsigned issues - extra test for signed/unsigned issues

+++ src/main/java/org/it4y/jni/linux/socket.java
@@ -11,11 +11,11 @@ package org.it4y.jni.linux;
  */
 //from include<bits/socket.h>
 public final class socket {
-    public final static int AF_UNSPEC=0;
-    public final static int AF_LOCAL=1;
-    public final static int AF_UNIX=AF_LOCAL;
-    public final static int AF_FILE=AF_LOCAL;
-    public final static int AF_INET=2;
-    public final static int AF_INET6=10;
+    public final static short AF_UNSPEC=0;
+    public final static short AF_LOCAL=1;
+    public final static short AF_UNIX=AF_LOCAL;
+    public final static short AF_FILE=AF_LOCAL;
+    public final static short AF_INET=2;
+    public final static short AF_INET6=10;
 
 }
